backend/requirements.txt
# Django core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async tasks
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Stripe payments
stripe==7.10.0

# Email
sendgrid==6.11.0

# PDF generation
reportlab==4.0.8
WeasyPrint==60.2

# QR codes
qrcode==7.4.2
Pillow==10.2.0

# Storage 
boto3==1.34.14
django-storages==1.14.2

# Security
python-dotenv==1.0.0
cryptography==41.0.7

# Monitoring
sentry-sdk==1.39.1

# WSGI server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Load testing
locust==2.20.1
